requirements.txt
python-dotenv>=1.0.0
PyAutoGUI>=0.9.54
pynput>=1.7.6
SpeechRecognition>=3.10.0
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0
pyaudio>=0.2.13
python-vlc>=3.0.18122